+++ arduino-pico/ToyStacky/TS-core-processDefaultPush.h
@@ -24,14 +24,14 @@ bool processDefaultPush(Machine* vm, char* token, char* retstr, StrackMeta* meta
 
 	} else {
 		//token is a real or complex number or string literal
-		double dbl;
+		long double dbl;
 		char rlc = strIsRLC(token);
 		bool s2c = stringToComplex(token, &c);
 		bool s2d = stringToDouble(token, &dbl);
 		//printf("processDefaultPush: A: token = %s\n", token);
 		if (s2c || s2d) {
 			if (s2d) {
-				//token is an RLC type double -- 1.234e3l
+				//token is an RLC type long double -- 1.234e3l
 				doubleToString(dbl, retstr, vm->precision, vm->notationStr);
 				char RLCstr[2];
 				RLCstr[1] = '\0';
